+++ metadata.rb
@@ -4,7 +4,9 @@ maintainer_email  'tsmith84@gmail.com'
 license           'Apache 2.0'
 description       'Installs and configures Nagios NRPE client'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           '1.5.3'
+version           '1.5.5'
+issues_url       'https://github.com/schubergphilis/nrpe/issues' if respond_to?(:issues_url)
+source_url       'https://github.com/schubergphilis/nrpe' if respond_to?(:source_url)
 
 recipe 'default', 'Installs and configures a nrpe client'
 %w(build-essential yum-epel).each do |cb|